Implemented ingestion by moving files.
Issue: Fixes #19

+++ fpi/tests/environment.py
@@ -2,6 +2,7 @@
 
 import os
 import os.path
+import shutil
 
 
 def remove_catalog(context):
@@ -25,14 +26,24 @@ def remove_tree(fselement):
 def before_scenario(context, scenario):
     """Execute before each scenario."""
     remove_catalog(context)
+    if "move" in scenario.feature.tags:
+        for d in ['data/catalog', 'data/originals']:
+            if os.path.isdir(d):
+                remove_tree(d)
+        shutil.copytree('data/samples/DCIM/100FPIAM', 'data/originals')
 
 
 def before_feature(context, feature):
     """Execute befora a feature is executed."""
     if "copy" in feature.tags:
-        remove_tree('data/catalog')
+        if os.path.isdir('data/catalog'):
+            remove_tree('data/catalog')
 
 
 def after_feature(context, feature):
     """Execute after a feature has been executed."""
     remove_catalog(context)
+    if 'copy' in feature.tags or 'move' in feature.tags:
+        for d in ['data/catalog', 'data/originals']:
+            if os.path.isdir(d):
+                remove_tree(d)
